Fixing signedness conversion warnings in libevmasm

+++ libevmasm/PeepholeOptimiser.cpp
@@ -84,7 +84,7 @@ struct SimplePeepholeOptimizerMethod
 	{
 		if (
 			_state.i + WindowSize <= _state.items.size() &&
-			ApplyRule<Method, WindowSize>::applyRule(_state.items.begin() + _state.i, _state.out)
+			ApplyRule<Method, WindowSize>::applyRule(_state.items.begin() + static_cast<ptrdiff_t>(_state.i), _state.out)
 		)
 		{
 			_state.i += WindowSize;
@@ -303,7 +303,7 @@ struct UnreachableCode
 {
 	static bool apply(OptimiserState& _state)
 	{
-		auto it = _state.items.begin() + _state.i;
+		auto it = _state.items.begin() + static_cast<ptrdiff_t>(_state.i);
 		auto end = _state.items.end();
 		if (it == end)
 			return false;
@@ -317,13 +317,13 @@ struct UnreachableCode
 		)
 			return false;
 
-		size_t i = 1;
+		ptrdiff_t i = 1;
 		while (it + i != end && it[i].type() != Tag)
 			i++;
 		if (i > 1)
 		{
 			*_state.out = it[0];
-			_state.i += i;
+			_state.i += static_cast<size_t>(i);
 			return true;
 		}
 		else
@@ -345,7 +345,7 @@ void applyMethods(OptimiserState& _state, Method, OtherMethods... _other)
 
 size_t numberOfPops(AssemblyItems const& _items)
 {
-	return std::count(_items.begin(), _items.end(), Instruction::POP);
+	return static_cast<size_t>(std::count(_items.begin(), _items.end(), Instruction::POP));
 }
 
 }
